Code cleanup

+++ include/apps/games/simon_game.h
@@ -25,7 +25,7 @@ class OswAppSimonGame  : public OswApp {
     std::vector<uint8_t> playerOrder;
 
     uint8_t score = 1;
-    unsigned long prevTime = 0;
+    uint32_t prevTime = 0;
     uint8_t printIdx = 0;
 
     bool gameRunning = false;
